delete acsses addet

+++ www/cms2/admin/includes/view_all_users.php
@@ -87,10 +87,16 @@ if (isset($_GET['change_to_sub'])){
 
 
 if (isset($_GET['delete'])){
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-    $delete_query =mysqli_query($connect,$query);
-    header("Location:users.php");
+    if (isset($_SESSION['user_role'])){
+        if($_SESSION['user_role']=='admin'){
+            $the_user_id = $_GET['delete'];
+            $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+            $delete_query =mysqli_query($connect,$query);
+            header("Location:users.php");
+        }
+
+    }
+
 }
